Added Sign Up Functionality
Added modal based on examples on Bootstrap to the nav bar called sign up
When clicked, it takes an eamail/un, password and confirm password.
Clicking the Sign Up button submits it to the server in a POST request.
The auth/index.js file sets the event handler, while events.js passes
the data and api.js contains the AJAX request to the server.
UI.js contains success messages and uses jquery to write to the DOM
that the person is signed up.