Finished Sign In Functionality
Added Sign in Modal functionality to request email and pw.  If correct,
user is signed in as player 1 if no players or player 2 if no players.
If there are already two players, an error is thrown.  The sign in uses
POST to connect to the server and AJAX to make the call.