Fix bugs and added Sign Out functionality
Added sign out functionality via a modal that shows only when a user
is signed in.  Allows two users to be signed in at once via the player
object.  Fixed a bug where sign out was not working because of a typo
and fixed another typo bug in the sign in success message